Color/Terrain/Brush/Lighting
Updating the colors. Lightening up the default object color. Pushing the
terrain further back in the scene and darkening it. Added in the updated
brush and fixed the pixelation issue. Adjusted the scene lighting and
shadows so everything looks a little better and removed the lines using
bias and normal bias under lighting.